Separate out tests for open_run() in new file
